Create a gt table theme using brand colors

Apply brand colors to a gt table.

Usage

theme_brand_gt(table, brand = NULL, background = NULL, foreground = NULL)

Arguments

table

A gt table object to theme.

brand

One of:

NULL (default): Automatically detect and read a _brand.yml file

A path to a brand.yml file or directory containing _brand.yml

A brand object (as returned by read_brand_yml() or as_brand_yml())

FALSE: Don't use a brand file; explicit colors must be provided

background

The background color, defaults to brand.color.background. If provided directly, this value can be a valid R color or the name of a color in brand.color or brand.color.palette.

foreground

The foreground color, defaults to brand.color.foreground. If provided directly, this value can be a valid R color or the name of a color in brand.color or brand.color.palette.

Value

Returns a themed gt table object.

See also

Examples

brand <- as_brand_yml('
color:
  palette:
    black: "#1A1A1A"
    white: "#F9F9F9"
    orange: "#FF6F20"
  foreground: black
  background: white
  primary: orange')

library(gt)
theme_brand_gt(
  gt(head(palmerpenguins::penguins)),
  brand
)

  

species
      	island
      	bill_length_mm
      	bill_depth_mm
      	flipper_length_mm
      	body_mass_g
      	sex
      	year
    

Adelie
	Torgersen
	39.1
	18.7
	181
	3750
	male
	2007

Adelie
	Torgersen
	39.5
	17.4
	186
	3800
	female
	2007

Adelie
	Torgersen
	40.3
	18.0
	195
	3250
	female
	2007

Adelie
	Torgersen
	NA
	NA
	NA
	NA
	NA
	2007

Adelie
	Torgersen
	36.7
	19.3
	193
	3450
	female
	2007

Adelie
	Torgersen
	39.3
	20.6
	190
	3650
	male
	2007

if (FALSE) { # rlang::is_installed("gt") && getRversion() < "4.5"
brand <- as_brand_yml('
color:
  palette:
    black: "#1A1A1A"
    white: "#F9F9F9"
    orange: "#FF6F20"
  foreground: black
  background: white
  primary: orange')

library(gt)
theme_brand_gt(
  gt(head(mtcars)),
  brand
)
}