Create a flextable theme using brand colors

Apply brand colors to a flextable table.

Usage

theme_brand_flextable(
  table,
  brand = NULL,
  background = NULL,
  foreground = NULL
)

Arguments

table

A flextable object to theme.

brand

One of:

NULL (default): Automatically detect and read a _brand.yml file

A path to a brand.yml file or directory containing _brand.yml

A brand object (as returned by read_brand_yml() or as_brand_yml())

FALSE: Don't use a brand file; explicit colors must be provided

background

The background color, defaults to brand.color.background. If provided directly, this value can be a valid R color or the name of a color in brand.color or brand.color.palette.

foreground

The foreground color, defaults to brand.color.foreground. If provided directly, this value can be a valid R color or the name of a color in brand.color or brand.color.palette.

Value

Returns a themed flextable object.

See also

Examples

brand <- as_brand_yml('
color:
  palette:
    black: "#1A1A1A"
    white: "#F9F9F9"
    orange: "#FF6F20"
  foreground: black
  background: white
  primary: orange')

library(flextable)
theme_brand_flextable(
  flextable(head(palmerpenguins::penguins)),
  brand
)

species

island

bill_length_mm

bill_depth_mm

flipper_length_mm

body_mass_g

sex

year

Adelie

Torgersen

39.1

18.7

181

3,750

male

2,007

Adelie

Torgersen

39.5

17.4

186

3,800

female

2,007

Adelie

Torgersen

40.3

18.0

195

3,250

female

2,007

Adelie

Torgersen

2,007

Adelie

Torgersen

36.7

19.3

193

3,450

female

2,007

Adelie

Torgersen

39.3

20.6

190

3,650

male

2,007

if (FALSE) { # rlang::is_installed("flextable") && getRversion() < "4.5"
brand <- as_brand_yml('
color:
  palette:
    black: "#1A1A1A"
    white: "#F9F9F9"
    orange: "#FF6F20"
  foreground: black
  background: white
  primary: orange')

library(flextable)
theme_brand_flextable(
  flextable(head(mtcars)),
  brand
)
}